correct overridden spelling
